+++ packages/convo-api-py/.pre-commit-config.yaml
@@ -1,0 +1,35 @@
+files: ^packages/convo-api-py/
+repos:
+-   repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v4.5.0
+    hooks:
+    -   id: check-merge-conflict
+    -   id: check-toml
+    -   id: check-yaml
+    -   id: check-json
+    -   id: end-of-file-fixer
+    -   id: trailing-whitespace
+    -   id: check-added-large-files
+    -   id: detect-private-key
+-   repo: https://github.com/pycqa/flake8
+    rev: 6.1.0
+    hooks:
+    -   id: flake8
+        args:
+          - "--max-line-length=120"
+          - "--max-complexity=18"
+          - "--select=B,C,E,F,W,T4,B9"
+          - "--ignore=W503,E203"
+          - "--per-file-ignores=__init__.py:F401"
+-   repo: https://github.com/pycqa/isort
+    rev: 5.12.0
+    hooks:
+    -   id: isort
+        args:
+          - "--multi-line=3"
+          - "--use-parentheses"
+          - "--trailing-comma"
+-   repo: https://github.com/psf/black-pre-commit-mirror
+    rev: 23.12.0
+    hooks:
+    -   id: black
